azure-ml-pipeline/src/pipeline-requirements.txt
azureml-sdk
azureml-pipeline-core 
azureml-pipeline-steps 
pandas 
requests
xgboost
sklearn
pyarrow